+++ bindings/python/parsers/srdf.cpp
@@ -4,6 +4,7 @@
 
 #include "pinocchio/parsers/srdf.hpp"
 #include "pinocchio/bindings/python/parsers/srdf.hpp"
+#include "pinocchio/bindings/python/utils/path.hpp"
 
 #include <boost/python.hpp>
 
@@ -14,6 +15,21 @@ namespace pinocchio
 
     namespace bp = boost::python;
 
+    void removeCollisionPairs(
+      const Model & model,
+      GeometryModel & geom_model,
+      const bp::object & filename,
+      const bool verbose = false)
+    {
+      pinocchio::srdf::removeCollisionPairs(model, geom_model, path(filename), verbose);
+    }
+
+    void loadReferenceConfigurations(
+      Model & model, const bp::object & filename, const bool verbose = false)
+    {
+      pinocchio::srdf::loadReferenceConfigurations(model, path(filename), verbose);
+    }
+
     void loadReferenceConfigurationsFromXML(
       Model & model, const std::string & xmlStream, const bool verbose = false)
     {
@@ -21,13 +37,18 @@ namespace pinocchio
       pinocchio::srdf::loadReferenceConfigurationsFromXML(model, iss, verbose);
     }
 
+    bool loadRotorParameters(Model & model, const bp::object & filename, const bool verbose = false)
+    {
+      return pinocchio::srdf::loadRotorParameters(model, path(filename), verbose);
+    }
+
     void exposeSRDFParser()
     {
 
       bp::def(
         "removeCollisionPairs",
-        static_cast<void (*)(const Model &, GeometryModel &, const std::string &, const bool)>(
-          &srdf::removeCollisionPairs),
+        static_cast<void (*)(const Model &, GeometryModel &, const bp::object &, const bool)>(
+          &removeCollisionPairs),
         (bp::arg("model"), bp::arg("geom_model"), bp::arg("srdf_filename"),
          bp::arg("verbose") = false),
         "Parse an SRDF file in order to remove some collision pairs for a specific GeometryModel.\n"
@@ -55,8 +76,8 @@ namespace pinocchio
 
       bp::def(
         "loadReferenceConfigurations",
-        static_cast<void (*)(Model &, const std::string &, const bool)>(
-          &srdf::loadReferenceConfigurations),
+        static_cast<void (*)(Model &, const bp::object &, const bool)>(
+          &loadReferenceConfigurations),
         (bp::arg("model"), bp::arg("srdf_filename"), bp::arg("verbose") = false),
         "Retrieve all the reference configurations of a given model from the SRDF file.\n"
         "Parameters:\n"
@@ -67,7 +88,7 @@ namespace pinocchio
       bp::def(
         "loadReferenceConfigurationsFromXML",
         static_cast<void (*)(Model &, const std::string &, const bool)>(
-          &srdf::loadReferenceConfigurations),
+          &loadReferenceConfigurationsFromXML),
         (bp::arg("model"), bp::arg("srdf_xml_stream"), bp::arg("verbose") = false),
         "Retrieve all the reference configurations of a given model from the SRDF file.\n"
         "Parameters:\n"
@@ -78,7 +99,7 @@ namespace pinocchio
 
       bp::def(
         "loadRotorParameters",
-        static_cast<bool (*)(Model &, const std::string &, const bool)>(&srdf::loadRotorParameters),
+        static_cast<bool (*)(Model &, const bp::object &, const bool)>(&loadRotorParameters),
         (bp::arg("model"), bp::arg("srdf_filename"), bp::arg("verbose") = false),
         "Load the rotor parameters of a given model from a SRDF file.\n"
         "Results are stored in model.rotorInertia and model.rotorGearRatio."