+++ examples/collision-with-point-clouds.py
@@ -4,7 +4,7 @@
 # pip install --user meshcat
 
 import sys
-from os.path import abspath, dirname, join
+from pathlib import Path
 
 import hppfcl as fcl
 import numpy as np
@@ -13,12 +13,12 @@ from pinocchio.visualize import MeshcatVisualizer
 
 # Load the URDF model.
 # Conversion with str seems to be necessary when executing this file with ipython
-pinocchio_model_dir = join(dirname(dirname(str(abspath(__file__)))), "models")
+pinocchio_model_dir = Path(__file__).parent.parent / "models"
 
-model_path = join(pinocchio_model_dir, "example-robot-data/robots")
+model_path = pinocchio_model_dir / "example-robot-data/robots"
 mesh_dir = pinocchio_model_dir
 urdf_filename = "panda.urdf"
-urdf_model_path = join(join(model_path, "panda_description/urdf"), urdf_filename)
+urdf_model_path = model_path / "panda_description/urdf" / urdf_filename
 
 model, collision_model, visual_model = pin.buildModelsFromUrdf(
     urdf_model_path, mesh_dir