+++ examples/overview-urdf.py
@@ -1,16 +1,15 @@
-from os.path import abspath, dirname, join
+from pathlib import Path
 from sys import argv
 
 import pinocchio
 
 # This path refers to Pinocchio source code but you can define your own directory here.
-pinocchio_model_dir = join(dirname(dirname(str(abspath(__file__)))), "models")
+pinocchio_model_dir = Path(__file__).parent.parent / "models"
 
 # You should change here to set up your own URDF file or just pass it as an argument of
 # this example.
 urdf_filename = (
-    pinocchio_model_dir
-    + "/example-robot-data/robots/ur_description/urdf/ur5_robot.urdf"
+    pinocchio_model_dir / "example-robot-data/robots/ur_description/urdf/ur5_robot.urdf"
     if len(argv) < 2
     else argv[1]
 )